Added helper functions

+++ RTC.h
@@ -15,7 +15,7 @@ void RTC_SetHour(uint8_t inputHour);
 uint8_t RTC_GetHour(void);
 void RTC_SetMinute(uint8_t inputMin);
 uint8_t RTC_GetMinute(void);
-uint16_t RTC_GetCurrentTime(void);
+double RTC_GetCurrentTime(void);
 uint16_t RTC_GetCurrentDay(void);
 void RTC_EnableInterrupt(void);
 void RTC_DisableInterrupt(void);
